Added exception handling

+++ src/main/java/com/medico/webapp/services/ProductService.java
@@ -1,11 +1,17 @@
 package com.medico.webapp.services;
 
+import com.medico.webapp.dto.Response;
 import com.medico.webapp.entity.Product;
+import com.medico.webapp.exception.BadRequest;
+import com.medico.webapp.exception.NotFound;
 import com.medico.webapp.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -13,20 +19,55 @@ public class ProductService {
   @Autowired
   private ProductRepository productRepository;
 
-  public List<Product> getAll() {
-    return productRepository.findAll();
+  public ResponseEntity<Response> getAll() {
+    List<Product> list = productRepository.findAll();
+    return new ResponseEntity<>(
+      new Response("Products are found successfully", list),
+      HttpStatus.OK
+    );
   }
 
-  public Product save(Product product) {
-    return  productRepository.save(product);
+  public ResponseEntity<Response> save(Product product) {
+    try {
+      Product savedProduct  = productRepository.save(product);
+      return new ResponseEntity<>(
+        new Response("Product is created successfully", savedProduct),
+        HttpStatus.CREATED
+      );
+    } catch (Exception ex) {
+      throw new BadRequest("Product details are not correct");
+    }
   }
 
-  public void delete(String id) {
-    productRepository.deleteById(id);
+  public ResponseEntity<Response> update(Product product) {
+    Optional<Product> fetchProduct = productRepository.findById(product.getId());
+    if(fetchProduct.isPresent()) {
+      Product updatedProduct = productRepository.save(product);
+      return new ResponseEntity<>(
+        new Response("Product is updated successfully", updatedProduct),
+        HttpStatus.OK
+      );
+    } else{
+      throw new NotFound("Product not found: " + product.getId());
+    }
   }
 
-  public Product getById(String id) {
-    return productRepository.findById(id)
-      .orElseThrow(() -> new RuntimeException("Product not found: " + id));
+  public ResponseEntity<Response> delete(String id) {
+    Optional<Product> product = productRepository.findById(id);
+    if(product.isPresent()) {
+      productRepository.deleteById(id);
+      return new ResponseEntity<>(new Response("Product is deleted successfully : " + id), HttpStatus.OK);
+    } else{
+      throw new NotFound("Product not found: " + id);
+    }
+  }
+
+  public ResponseEntity<Response> getById(String id) {
+    Optional<Product> product = productRepository.findById(id);
+    if(product.isPresent()) {
+      return new ResponseEntity<>(new Response("Product is found successfully", product), HttpStatus.OK);
+    } else{
+      throw new NotFound("Product not found: " + id);
+    }
   }
 }
